veux: Declare support for blur
The Adreno 660 should be more than capable of handling this even though
we know that there is a slight performance hit. However, this shouldn't
matter all that much as blur isn't really used in a lot of places.

This also enables a runtime toggle in Developer Options that disables it
in case someone doesn't want it.

Change-Id: I20d9599bd76d11ceb685875268a44b6363576666
Reviewed-on: https://review.statixos.com/c/android_device_asus_sake/+/7999
Reviewed-by: Sourajit Karmakar <sourajit@live.com>
Tested-by: Sourajit Karmakar <sourajit@live.com>